requirements.txt
alembic==1.12.1
Babel==2.14.0
blinker==1.6.3
click==8.1.8
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.5
flask-babel==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.5
Mako==1.2.4
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==0.21.1
pytz==2024.2
SQLAlchemy==2.0.36
typing_extensions==4.7.1
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
